--- frontend/src/main/resources/webapp/src/app/main/events/events-detail/events-detail.component.html
@@ -1,37 +1,0 @@
-<div fxLayout="column" fxLayoutAlign="space-evenly stretch" class="event-detail" fxFill>
-
-  <div fxLayout="row" fxLayoutAlign="space-between center" fxLayoutGap="24px">
-    <div fxLayout="column" fxLayoutAlign="center center">
-      <div class="image" style="margin-bottom: 10px;">
-        <img style="height: 150px; width: 175px;" src="assets/images/events/{{data.event_name}}.jpg">
-      </div>
-      <span class="event-name" style="text-align: center; font-size: large;"><b>{{ data.event_name }}</b></span>
-      <mat-divider></mat-divider>
-    </div>
-    <div fxLayout="column" fxFlex="75" class="event-data">
-      <span class="event-name" style="text-align: center; font-size: largeR;"><b>{{'INFO' | oTranslate }} </b></span>
-      <span class="event-address" style="margin-top: 10px;">
-        <mat-icon>home</mat-icon>{{ data.event_address }} {{ data.event_region }}
-      </span>
-      <span class="event_date_time">
-        <mat-icon>star</mat-icon>{{ data.event_date_time |  date:'dd/MM/yyyy'}}
-      </span>
-
-      <o-list [static-data]="getEventResult()" title="" quick-filter="false"
-      refresh-button="false" insert-button="false" insert-button-floatable="false" delete-button="false"
-      selectable="true" dense="false" detail-button-in-row="false" edit-button-in-row="false" detail-mode="none">
-      <o-list-item *ngFor="let event of eventResult">
-        <p class=item-hover [routerLink]="['/main/event', event.event_id]" [queryParams]="{id: event.event_id}"
-          href="#">{{ event.band_name}}</p>
-      </o-list-item>
-    </o-list>
-     
-
-      <span class="event_buy_tickets" style="margin-top: 10px; align-items: center;">
-        <o-button attr="flat" (click)="openURL(data)" type="FLAT" label="{{'BUYTICKETS' | oTranslate }}" layout-padding>
-        </o-button>
-
-      </span>
-
-    </div>
-  </div>
