+++ frontend/src/main/resources/webapp/src/app/main/events/events-routing.module.ts
@@ -1,15 +1,15 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { EventsHomeComponent } from './events-home/events-home.component';
-import { EventsDetailComponent } from './events-detail/events-detail.component';
+import { SharedModule } from 'app/shared/shared.module';
 
 const routes: Routes = [{
   path:'',
   component: EventsHomeComponent
 },
 {
-  path:'/EVENT_ID',
-  component: EventsDetailComponent
+  path:':event_id',
+  component: SharedModule
 }];
 
 @NgModule({
