+++ frontend/src/main/resources/webapp/src/app/main/events/events.module.ts
@@ -3,18 +3,24 @@ import { CommonModule } from '@angular/common';
 import { OntimizeWebModule } from "ontimize-web-ngx";
 import { EventsRoutingModule } from './events-routing.module';
 import { EventsHomeComponent } from './events-home/events-home.component';
-import { EventsDetailComponent } from './events-detail/events-detail.component';
+import { SharedModule } from 'app/shared/shared.module';
+import { EventsDetailComponent } from 'app/shared/events-detail/events-detail.component';
+
+
 
 @NgModule({
   imports: [
     CommonModule,
     OntimizeWebModule,
-    EventsRoutingModule
+    EventsRoutingModule,
+    SharedModule
   ],
   declarations: [
     EventsHomeComponent,
     EventsDetailComponent
-
+  ],
+  entryComponents: [
+    EventsDetailComponent
   ]
 })
 export class EventsModule { }
