+++ frontend/src/main/resources/webapp/src/app/main/home/home-routing.module.ts
@@ -1,17 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-
 import { HomeComponent } from './home.component';
-import { HomeeventsDetailComponent } from './homeevents-detail/homeevents-detail.component';
 
 const routes: Routes = [
   {
     path: '',
     component: HomeComponent
-  },
-  {
-    path:'/EVENT_ID',
-    component: HomeeventsDetailComponent
   }
 ];
 
@@ -20,3 +14,4 @@ const routes: Routes = [
   exports: [RouterModule]
 })
 export class HomeRoutingModule { }
+
