+++ frontend/src/main/resources/webapp/src/app/main/home/home.component.html
@@ -1,46 +1,42 @@
-
 <div fxLayout="column" fxFlex="1 1 auto" style="    background: repeating-linear-gradient( 44deg, rgba(0, 0, 0, 0) 40px,
   rgba(0, 0, 0, 0) 48px, #e4e3e3 20px, #e4e3e3 49px);">
-        <div class="div_bandsEmph">
-  <o-grid #grid attr="BANDTABLE" title="DESTACADOS" service="bands" entity="getBandsEmph"
-    columns="band_name;CATEGORY" keys="band_id" query-rows="1" 
-    refresh-button="no" gutter-size="18px" detail-mode="none" quick-filter="false" style=" margin: 5% 5% 5% 5%; " class="o-grid-home">
-    <o-grid-item *ngFor="let item of grid.dataArray">
-      <div fxLayout="column" fxLayoutAlign="space-evenly center"
-        class="band-item mat-elevation-z1">
-        <div fxLayout="row" fxLayoutAlign="space-between end" class="image-container">
-          <mat-divider fxFlex></mat-divider>
-          <div class="image">
-            <img src="assets/images/bands/{{item.band_name}}.jpg"> 
+  <div class="div_bandsEmph">
+    <o-grid #grid attr="BANDTABLE" title="DESTACADOS" service="bands" entity="getBandsEmph" columns="band_name;CATEGORY"
+      keys="band_id" query-rows="1" refresh-button="no" gutter-size="18px" detail-mode="none" quick-filter="false"
+      style=" margin: 5% 5% 5% 5%; " class="o-grid-home">
+      <o-grid-item *ngFor="let item of grid.dataArray">
+        <div (click)="openBandsDetail(item)" fxLayout="column" fxLayoutAlign="space-evenly center"
+          class="band-item mat-elevation-z1">
+          <div fxLayout="row" fxLayoutAlign="space-between end" class="image-container">
+            <mat-divider fxFlex></mat-divider>
+            <div class="image">
+              <img src="assets/images/bands/{{item.band_name}}.jpg">
+            </div>
+            <mat-divider fxFlex></mat-divider>
           </div>
-          <mat-divider fxFlex></mat-divider>
+
+          <div fxLayout="column" fxLayoutAlign="start center" class="band-data">
+            <span class=" band-name">{{ item.band_name }}</span>
+            <span>{{  item.CATEGORY  }}</span>
+
+          </div>
         </div>
 
-        <div fxLayout="column" fxLayoutAlign="start center" class="band-data">
-          <span class=" band-name">{{ item.band_name }}</span>
-          <span>{{  item.CATEGORY  }}</span>
+      </o-grid-item>
+    </o-grid>
+  </div>
 
-          </div>
-        </div>
- 
-    </o-grid-item> 
-  </o-grid>
+
+  <div class="div_eventsEmph">
+    <o-list #list attr="EVENTTABLE" service="events" title="PROXIMOS EVENTOS" entity="event" keys="event_id"
+      columns="event_name;event_date_time;event_region" refresh-button="true" quick-filter="false" insert-button="false"
+      delete-button="false" row-height="medium" detail-mode="none" refresh-button="no" gutter-size="0">
+      <o-list-item *ngFor="let row of list.dataArray">
+        <o-list-item-card (click)="openDetail(row)" #item title="{{row.event_name}}"
+          subtitle=" {{ row.event_region}}, {{row.event_date_time |  date:'dd/MM/yyyy'}} " show-image="true"
+          image="assets/images/events/{{row.event_name}}.jpg"> </o-list-item-card>
+
+      </o-list-item>
+    </o-list>
+  </div>
 </div>
-
-
-<div class="div_eventsEmph">
-  <o-list #list attr="EVENTTABLE" service="events" title="PROXIMOS EVENTOS" entity="event" columns="event_name;event_date_time;event_region"
-  refresh-button="true" quick-filter="false" insert-button="false" delete-button="false"
-  row-height="medium" detail-mode="none" refresh-button="no" gutter-size="0">
-   <o-list-item *ngFor="let row of list.dataArray" >
-      <o-list-item-card (click)="openDetail(row)"  #item title="{{row.event_name}}"  subtitle=" {{ row.event_region}}, {{row.event_date_time |  date:'dd/MM/yyyy'}} " show-image="true" image="assets/images/events/{{row.event_name}}.jpg" > </o-list-item-card>
-
-    </o-list-item>
-</o-list>
-</div> 
-  
-
-
-</div>
-
-
