+++ frontend/src/main/resources/webapp/src/app/main/home/home.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { MatDialog } from '@angular/material';
-import { HomeeventsDetailComponent } from './homeevents-detail/homeevents-detail.component';
+import { BandsDetailComponent } from '../../shared/bands-detail/bands-detail.component';
+import { EventsDetailComponent } from '../../shared/events-detail/events-detail.component';
+
 
 @Component({
   selector: 'home',
@@ -25,11 +27,20 @@ export class HomeComponent implements OnInit {
   }
 
   public openDetail(data: any): void {
-    this.dialog.open(HomeeventsDetailComponent, {
+    this.dialog.open(EventsDetailComponent, {
       height: '330px',
       width: '520px',
       data: data
     });
   }
+
+  public openBandsDetail(data: any): void {
+    this.dialog.open(BandsDetailComponent, {
+      height: '98%',
+      width: '80%',
+      data: data
+    });
+  }
+  
 
 }
