+++ frontend/src/main/resources/webapp/src/app/main/main.component.html
@@ -1,14 +1,13 @@
- <o-bar-menu id="upmenu">
+<o-bar-menu id="upmenu">
   <o-bar-menu-group title="OPTIONS">
     <o-bar-menu-group title="LANGUAGE">
       <o-locale-bar-menu-item locale="en" title="English"></o-locale-bar-menu-item>
       <o-locale-bar-menu-item locale="es" title="Español"></o-locale-bar-menu-item>
     </o-bar-menu-group>
     <o-bar-menu-separator></o-bar-menu-separator>
-    <o-bar-menu-item title="LOGOUT" icon="power_settings_new" route="/login" confirm="MESSAGES.CONFIRM_LOGOUT"></o-bar-menu-item>
+    <o-bar-menu-item title="LOGOUT" icon="power_settings_new" route="/login" confirm="MESSAGES.CONFIRM_LOGOUT">
+    </o-bar-menu-item>
   </o-bar-menu-group>
 </o-bar-menu>
 
-  <router-outlet></router-outlet>
-
-  
+<router-outlet></router-outlet>
