+++ frontend/src/main/resources/webapp/src/app/main/main.component.ts
@@ -1,6 +1,4 @@
 import { Component, OnInit, ViewEncapsulation, enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
 
 
 @Component({
