+++ frontend/src/main/resources/webapp/src/app/shared/bands-detail/bands-detail.component.ts
@@ -1,7 +1,6 @@
-import { Component, OnInit, Inject, ViewChild} from '@angular/core';
+import { Component, OnInit, Inject, ViewChild } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 import { MAT_DIALOG_DATA } from '@angular/material';
-
 
 
 @Component({
@@ -10,23 +9,14 @@ import { MAT_DIALOG_DATA } from '@angular/material';
   styleUrls: ['./bands-detail.component.scss']
 })
 
-@Component({
-  selector: 'ngbd-carousel',
-  templateUrl: './bands-detail.component.html'
-})
-
-
 export class BandsDetailComponent implements OnInit {
-
-
 
   constructor(
     @Inject(MAT_DIALOG_DATA) public data: any,
     protected sanitizer: DomSanitizer
-    
+
   ) { }
-  
- 
+
   videos: string[] = [
     'SkypZuY6ZvA',
     '_N5YIGT_R58',
@@ -34,15 +24,7 @@ export class BandsDetailComponent implements OnInit {
     'uelHwf8o7_U'
   ];
 
-
-  ngOnInit()  {}
-
-
-  images: string[] = [944, 1011, 984].map((n) => 'https://picsum.photos/id/${n}/900/500');
-  
-  ngOnInit() {
-    
-  }
+  ngOnInit() {}
 
 }
