+++ frontend/src/main/resources/webapp/src/app/shared/shared.module.ts
@@ -2,14 +2,17 @@ import { NgModule } from '@angular/core';
 import { OntimizeWebModule } from 'ontimize-web-ngx';
 import { CommonModule } from '@angular/common';
 
+
+
 @NgModule({
   imports: [
     OntimizeWebModule
   ],
   declarations: [
+    
   ],
   exports: [
-    CommonModule
+    CommonModule,
   ]
 })
 export class SharedModule { }
